requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
debugpy==1.6.3
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
executing==1.0.0
idna==3.3
ipykernel==6.15.2
ipython==8.4.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==23.2.1
requests==2.28.1
six==1.16.0
stack-data==0.5.0
tornado==6.2
traitlets==5.3.0
urllib3==1.26.12
wcwidth==0.2.5
yarg==0.1.9
